feat: Criação da rota de pagamento por boleto
