+++ app/routes/api_config.py
@@ -8,7 +8,8 @@ from .v1_endpoint import (
     imagem_usuario_routes, 
     endereco_usuario_routes, 
     avaliacao_produtos_routes, 
-    historico_compras_usuario_routes
+    historico_compras_usuario_routes,
+    pagamento_bol_routes
 )
 
 
@@ -24,4 +25,5 @@ api_router.include_router(produto_routes.router, prefix='/produtos', tags=['prod
 api_router.include_router(avaliacao_produtos_routes.router, prefix='/avaliacao-produtos', tags=['avaliacao-produtos'])
 api_router.include_router(vendedor_routes.router, prefix='/vendedor', tags=['vendedor'])
 api_router.include_router(historico_compras_usuario_routes.router, prefix='/historico-compras-usuario', tags=['historico-compras-usuario'])
+api_router.include_router(pagamento_bol_routes.router, prefix='/pagamento_boleto', tags=['pagamento_boleto'])
